Ensure metadata values are updated when modifying draft graph.
Add an upsert function for inserting and updating the object
for subject/predicate pair which are expected to have a single
associated value. Use this function when updating graph metadata
so that existing values are updated instead of appending new
values.

Update set-isPublic! to use the upsert function.

Add tests for the upsert functionality and to check that metadata
is updated when modifying the draft.