Cancel queries that fail to write results or complete within a timeout period.
Cancel long-running queries that take too long to either complete or write the
next query result. Executing SPARQL queries requires two threads - one to service
the client and one to asynchronously write query results to a shared buffer. If
the client disconnects, the writing thread will not discover the client connection
has closed until it tries to write into the buffer. This means queries that
take a long time to produce a result can continue executing and causing
contention for other queries. Apply two timeouts to all SPARQL queries - one for
the time to write the next result, and one for the entire query operation. If
either of these timeouts is exceeded the query should be cancelled, which
should cause the client connection to be closed.

SPARQL query operations are now executed by a monitor which encapsulates a
java ExecutorService. The monitor maintains a map of executing operations
along with the start time, query timeouts and the time the last result was
written. The monitor periodically checks all outstanding operations to find
those queries which have exceeded either timeout period. Any such operations
are cancelled by the monitor. All operations have an associated publish
function which the operation can use to report when it has written a result
to the client.

Operations are constructed and executed in three stages:
1. Create an operation on a monitor. This gives access to the publish
function to notify result writes.

2. Associate the query function with the operation - this separation is
required to allow the query operation to access the publish function
exposed by the operation.

3. Execute the operation on the associated monitor. This runs the operation
on the monitor's ExecutorService.