Remove 'monitor' concept from operations namespace.
Remove the concept of a monitor from drafter.operations and require the
executor service to be supplied when executing an operation. Creating
an operation now requires only the target atom the operation will be
tracked in, and the time source.

Use scheduled executor for the repeating task instead of explicit and thread.

The concept of a monitor has been replaced with a 'reaper' - reapers have
their own monitor task and require the operations atom to monitor for
timed out operations.