Replace concept of 'event' with 'timestamp'.
Operation status is calculated based on the time of the last write
event and the associated timeout periods. This is currently called
the last 'event' which may imply a compound structure rather than
a simple value. Remove all references to operation events and use
timestamp instead.

Shorten the line-length of all function docstrings in the operations
namespace.