Hard-code offset calculation for clocks.
Remove the offset-fn member from clocks and always use the + operator
for calculating the offset of a time and timespan element. The clock
parameter to offset is now not needed however it is kept in case
more sophisticated offset calculations are needed per clock in future.

Make the clock functions private to the operations namespace.