Move operation 'connect' and 'register' functionality into execute.
Clients no longer need to explictly connect operations to their underlying
functions since this is done in the execute method. Given the function
to execute there are four steps required to execute an operation:

1. Create the java.util.concurrent.Future instance to execute
2. Mark the operation as having been run so it cannot be re-submitted.
3. Register the future and the operation state in the operation atom
4. Execute the operation on the supplied executor.

1 and 2 can be done by registering a FutureTask in the promise associated
with the operation. If step 4 throws an exception then the submission
of the future on the executor failed. In this case the operation can
be considered failed and does not need to be monitored. Step 3 is simply
adding the promise -> state pair into the map.

All the above steps are now done directly in execute-operation so
connect and register can be removed. init-operation state is private and
only called from one place so it has been placed inline in execute-operation.