Apply timeout configuration to all SPARQL endpoints.
All SPARQL endpoints are now provided with their own timeout configuration
on construction. In order to calculate the timeout configuration, all
endpoint names need to be known up front so the configuration
parser can warn about unknown endpoint names found in the environment.
SPARQL endpoints are now 'specified' before being constructed - the
specification defines how to create the handlers for the query and
update routes for the endpoint.

On initialisation, the specification for all endpoints are collected
in a map and then used to derive the timeout configuration. These
are then combined to create the routes for all SPARQL endpoints.