Cancel timed-out SPARQL update operations.
Monitor SPARQL update operations and cancel any which have timed out.
Add a function in the operations namespace to register a Future
for cancellation directly instead of needing to create an 'operation'
instance. Wrap the SPARQL update in a Future which is executed on
the request thread and handles the thread interrupt if the Future
is cancelled in the event of timeout.