Rewrite all graphs in queries and results.
Rewrite all graph constants found in queries and results instead of
just those in graph position. If a graph is a subject/predicate/object
of a triple, it should still be re-written.

Rewrite all URIs found in ValueConstants in the query AST - these can
appear in the predicate for a FILTER clause.