Include class name, stack trace and cause in serialised exceptions.
Add the fully-qualified Java class name of exception objects when
serialising to JSON. Include a vector containing the string representation
of each stack frame. If an exception has a non-null cause, include the
serialised representation of the cause - this creates a nested tree
of exception documents if causes are present in the top-level exception.