Execute SPARQL updates through the jobs queue.
Execute all SPARQL updates using the jobs queue with a sync-write
priority. This means that updates requests will return a 503
'service unavailable' response if an exclusive write operation is in
progress. If there is no exclusive write operation, the update job is
queued and the request blocked until the result is available.

Extract the common request logic for update endpoints into a single
function.

Update SPARQL update tests to use the test database and start/stop the
write queue writer.