Create custom SPARQL client for use in Drafter.
Stardog operations require that the 'include inferred' property be set to
false to work correctly. Failure to do this causes queries to return
results not in the database, and deletes to silently fail. Create
subclasses of SPARQLRepository and SPARQLConnection which ensure that
this flag is set to false for all query and update operations.

Note that when adding statements through the repository, the base
SPARQLClient creates its own SPARQLUpdate command inline which
will have this flag set since it is the default. This does not appear
to cause issues with updates so the derived client does not override
commit to handle this case.