Migrate draft graphs using SPARQL MOVE query.
Use SPARQL MOVE query to migrate a draft graph to live. When using the
SPARQL repository, transactions accumulate statement inserts and deletes
in a SPARQL string which is submitted to the server when the transaction
is committed. The existing approach of manually copying all statements
in the draft graph is therefore inefficient over the SPARQL repository
and may result in 'out of memory' exceptions if the source graph is large.

Inside the SPARQLConnection class, only methods which add/remove statements
directly take part in the transaction - this means that any MOVE statement
executed inside the scope of begin/commit transaction calls will not actually
take part in the transaction. The migrate job now manually accumulates
the SPARQL statements required to migrate all draft graphs in a draftset
into a single statement and submits the compound update. This means the
explicit transaction can be removed.