Only visit prologue in top-level SPARQL queries when re-writing.
Subqueries in ARQ inherit the prologue from the top-level query. When
rewriting queries this prologue should only be copied into the top-level
definition and not any of the subqueries. Add a flag to the query visitor
to indicate if the current query is the top-level one, and avoid creating
a prologue for the re-written query if it is not.