Validate SPARQL queries syntactically before evaluation.
Attempt to validate SPARQL queries before submitting them for evaluation
by parsing them. If the parse fails, a 400 response is returned to the
client indicating the invalid query. This attempts to catch a common
class of errors which may occur when processing SPARQL queries to
give a more accurate response to the client. Delaying processing of
the query means that the 200 response headers will be returned to ring
before the body has been evaulated.

Move the parsing and rendering functionality from ARQ into its own
drafter.rdf.arq namesapce so it can be depended on by both
sparql-protocol and sparql-rewriting.