Separate synchronous and asynchronous job handling.
Move all HTTP response handling into the new drafter.responses
namespace. The write-scheduler namespace is now only concerned with
submitting jobs and waiting for their results. Jobs are now added to
the write queue using the queue-job function.

Remove submit-job! and replace it will two functions - submit-async-job!
and submit-sync-job! for async and sync jobs respectively. Async jobs
work as before and return a ring response describing the status of the
submit operation. submit-sync-job! requires an additional handler
argument to convert the job result into a ring response to be returned
to the client.

Ensure no jobs complete with a ring response - the async job handling
is responsible for constructing ring responses from job results. Jobs
should only submit a map representing the JSON result to send back to
the client.