Clone live graph in batches.
When appending data to a draft graph, clone the live graph in batches
to prevent a long-running clone operation.

Remove the clone step from append-data! as this is done within the
clone job. One of the tests draft_management_tests.clj depends on
the old clone behaviour in append-data! so create a function with
the required clone behaviour in the test namespace.