Look for metadata parameters in the query string and request body.
Search in the request body for metadata parameters in the metadata and
update routes in the drafts_api namespace. The 'params' map in the ring
request contains the merged map of query parameters and form parameters
so update the handler functions to search the params map instead of the
query-params map.