Create protocol for executing SPARQL queries.
Create a SparqlExecutor protocol for executing SPARQL queries. This
protocol allows process-sparql-query to execute queries and perform
content negotiation without any knowledge of the concrete backend
used to execute the queries.

Create an implementation of the SparqlExecutor which executes queries
through a sesame repository.

Create another implementation for sesame repositories which performs
query and result rewriting and move it out of process-sparql query.
The draft query endpoint now constructs this wrapper and provides it
the live->draft graph mapping.