Create default implementations of backend protocols.
Move the implementations of the backend protocols into the
backend.sesame-common namespace. Each protocol has a corresponding
map containing the mapping from the protocol member name to a fn
for the default implementation.

Update the existing backend record to implement the protocols using
extend and providing the default implementation maps. Different backends
will be able to override protocol methods by updating the key-value
pairs in the default implementation maps.