Move append-data! to the DraftManagement protocol.
Remove draft-management/append-data! and add it to the DraftManagement
protocol with the name append-data-batch!. This takes a sequence of
triples to add to a given draft graph. Rewrite append-data-to-graph-job
to use the protocol method rather than append-data! Rewrite the
tests to use the protocol and remove the old append-data! method.