Move delete-graph-job implementation into backend.sesame-common.
The current delete-graph-job implementation deletes the source graph
in batches in an effort to reduce latency for concurrent users since
DROP operations on sesame native are slow. Move the existing
implementation into the sesame-native namespace.

The method of deleting a single graph batch may differ between sesame
repository implmentations so add a new SesameBatchOperations protocol
and add a default implementation for the sesame native backend.