Move migrate-live! to the DraftManagement protocol.
Move migrate-live! to the DraftManagement protocol with the modification
that it takes a sequence of draft graphs to migrate. The existing
function now just calls the protocol method. Move the implementation
of migrate-graphs-to-live into the backend.sesame-common namespace.

Implement the ITripleWriteable protocol on the sesame native backend.
This is used by many of the tests which need to add some statements to
a draft and then migrate it to live.