Add protocol for backends that provide access to a sesame repository.
Add a ToRepository protocol which defines a ->sesame-repo method for
fetching a sesame Repository instance from a backend. Many of the
common backend functions are implemented around the sesame Repository
interface so this gives a uniform way of implementing them for sesame
backends. Implement the protocol on the Stardog and sesame native
backends.