Access backend instead of the sesame repository in tests.
Change test-common/make-backend to only return the backend instead of
the backend and associated sesame repository. Update the tests to
add/update/query from the backend directly.

The SPARQL result rewriting tests still have a small dependency on the
Sesame repository in that they expect queries prepared by the backend
to be sesame query types.