Change 'update metadata' operation on the DraftManagement protocol.
Rename append-graph-metadata! to append-metadata-to-graphs! and change
the graph parameter to take a sequence of graphs to append the
metadata to. Add an append-graph-metadata! function which wraps the
given graph into a single-element sequence and calls the new method.

This is because different backends may be able to perform the update
more efficiently across all graphs rather than updating each
graph individually.