Update graph metadata with a single query on SPARQL backends.
Sesame native submits multiple 'upsert' statements for each metadata
pair being updated for each graph in append-metadata-to-graphs! This
is inefficient for remote SPARQL backends which could submitting
a large number of small updates. Create a query to perform the update
in a single INSERT/DELETE statement and submit that on SPARQL repository
backends (SESAME_SPARQL and STARDOG_SESAME).