Extract repeated scheduling logic for batched jobs.
Batched jobs contain repeated logic for extracting the next batch
of work to execute and scheduling the continuation job. Extract this
logic into one which schedules a sequence of 'joblets' in order until
all have been executed. This means the batching logic can be
separated from the scheduling logic by constructing a joblet for
each batch and scheduling them all under a parent job.

Rewrite append-data-in-batches to construct a sequence of joblets
for the input data and metadata before submitting the sequence for
execution.