Add route to append quad data to a draftset.
Add a route at /draftset/:draftset-id/data to allow quad data to be
appended to a draftest. Any graphs in the quad data without a
corresponding draft graph in the draftset will be created on demand.

The append job takes batches of quads from the incoming data,
groups them by graph and then appends each graph batch into the
draftset.