Ensure client provides a valid RDF quad format when appending to draftset.
When appending quads to a draftset, ensure the client provides a content
type for either the request or file part containing the data. Validate
this content type can be resolved to a sesame RDF format which supports
quad data. If any of these conditions are not met, return a client
error response.