Add route to get all draftsets.
Add route to get information for all draftsets. Like constructing
the information for a single draftset, the data is fetched in two
stages. First the data associated with the draftset itself (title,
description etc.) and then the live->draft graph mapping is fetched
and converted into the mapping for the "data" key. Rather than
repeating this process for each draftset, all the dataset info and
their graph mappings are fetched at once and then reconstructed.