Move 'all quads' query to the backend protocol.
Add an all-quads-query method to the SparqlExecutor protocol. There is
no standard way to fetch a subset of quads using SPARQL so this is
left up to each individual backend to implement. One approach is to
issue a select (?s ?p ?o ?g) query and then map each returned set of
bindings to a quad in the results. The current implementation for
the Sesame backend uses this approach, but the RepositoryConnection
class has a getStatements method which could possibly be used instead.