Add route to delete a set of quads from a draftset.
Add a route which accepts a serialised set of quads to delete from
a draftset.

Add a new backend protocol which removes a set of statements from the
underlying datastore. Add an implementation for the
RewritingSesameSparqlExecutor which rewrites the incoming statements
according to the live->draft graph mapping and submits them to its
inner backend for deleting.