Exract common logic from draftset API handlers.
Many of the draftset API handlers contain common logic for checking
whether the target draftset exists, and returning a 404 response if
it does not. Extract this logic into a middleware handler which
contains the check. Since this approach results in the Ring handler
function being created explicitly, call the compojure make-route
function instead of using the GET/PUT/DELETE macros for defining
handlers.