Add middleware for validating the format of serialised RDF files.
Add a Ring middleware for requests expected to contain file parts
containing serialised RDF data. The middleware checks a content
type for the data is provided on the request and also validated the
content type maps to a know RDF format. If so, it associated this
information with the request and invokes the inner handler, otherwise
a client error response is returned.