Keep draft graphs for managed graphs deleted inside a draftset.
Draft graphs inside a draftset represent the desired state of the
corresponding live graph after a publish operation. Any managed
graphs without a corresponding draft graph in the draftset are not
affected by a publish operation. Therefore, a deleted graph should
still have a corresponding (but empty) draft graph inside a draftset.
When the draftset is published, these empty draft graphs will result
in deletions of the corresponding live graphs.

Change rdf.draftset_management/delete-draftset-graph! to ensure that
a draft graph exists for a managed graph inside a draftset if the
corresponding live graph exists. If no live graph exists, this is
considered a no-op since draftset graph operations are considered
to be updates to live at the point they were made.