Lazily copy live graphs when deleting statements from a draftset.
When deleting statements from a draftset, lazily copy the live graph
into the draftset before performing the deletion. If the graph does
not exist in live then do nothing and ignore the current batch (the
statements to delete are grouped by graph during the delete operation).
This behaviour is consistent with the behaviour for deleting entire
graphs from a draftset where the operation is considered to act on
the state of live at the time the request is processed.