Make the draftest API route handler visible to all tests.
Add a dynamic var to the draftsets API tests and bind it to a new handler
before every tests is executed. Add a wrapper function to call the
bound hander to the test namespace. Remove the top-level let binding
to a new handler function in each of the test functions.