Use better-errors stuff in lib-swirrl-server
Change drafter to use encode-error and ex-swirrl for raising exceptions
and ring json error responses.

This includes some changes to the way that the async jobs middleware
works, as it now raises an exception when it 503's because of the write
lock.  Hence the tests have been updated to catch the exception and then
test that it encodes as the right ring response.