Create authentication Ring middleware using buddy-auth.
Create a Ring middleware which authenticates incoming requests using
HTTP Basic authentication. The incoming credentials are used to lookup
a user in a given user store and authenticates the credentials against
the user's API key if the user exists. The middleware adds an :identity
key to the request map which is mapped to the corresponding
drafter.user/User record if authentication succeeded, otherwise it is
nil.