Associate the user with a draftset on creation.
When creating a draftset, associate a 'created by' statement between
the draftset and the user. Initially set the creating user as the
current owner through a 'has owner' statement.

The create draftset route now always expects a valid user to be
present on the incoming request so update the draftset API tests
to ensure one is always provided.