Specify the HTTP API for the draftset lifecycle workflow.
Change the 'submit' and 'reject' routes on the draftset API to be 'offer'
and 'return' respectively. In the 'offer' route, a user gives up ownership
of a draftset and makes it available to users with a given role. Users
in that role can then take ownership of the draftset and make any
changes permitted by their privileges. Such users can then return
ownership of the draftset back to the original owner (i.e the user
who first created it).