Work around what looks like a bug in lein profiles and uberjars.
We've had this before, basically profiles for the uberjar task aren't
properly merged; so both jenins and devs need to manage their
credentials in the same kinda way or things will fail.

We've probably done & undone this change a million times; this is the
million-and-oneth!

You may need to generate uberjars etc with:

env LEIN_USERNAME=AKIA... LEIN_PASSPHRASE=blahblah lein uberjar