Implement 'claim draftset' route.
Add a route to allow users in the specified role to claim ownership
of draftsets which have previously been offered. If the draftset is
available to be claimed and the user is in the specified role, the
claiming user will become the new owner of the draftset. If the
draftset is not on offer, or the user does not have sufficient
privileges to claim the draftset, a 403 Forbidden response is returned.
If a user attempts to claim a draftset they already own, no changes
are made and the request succeeds.