Change the set of valid user roles to an ordered list.
Change drafter.user/roles from an unordered set to an ordered collection
from least to most-privileged role.

Add a seq-contains? function to drafter.util which tests whether a
sequence contains a given value. Note that clojure.core/contains?
tests whether get would return a value for the input collection,
which for vectors means checking whether the index is in range.

Add an is-known-role? function to drafter.user to centralise all
role validation into the user namespace.