Optimistically claim draftset by user.
claim-draftset! currently checks the claiming user is able to claim
the draftset and update the draftset state if they are permitted. If
the check fails then an error is returned. The claim process is subject
to a race condition if the draftset state changes between the check and
the update to assign the claiming user as the owner.

Add a try-claim-draftset! function which performs the check and update
in a single statement. Modify claim-draftset! to attempt the update and
then check the subsequent state of the draftset. If the claiming user
is the new owner of the draftset then claim succeeded, otherwise an
error response is returned.