Add draftset module.
Add a draftset module which defines an in-memory representation of
a draftset. Move the existing DraftsetRef protocol and DraftsetId
and DraftsetURI records from draftset-management into the new module,
and fix existing references.

Add a permitted-draftset-operations function to drafter.user which
returns all the operations a user can carry out on a draftset in its
current state.