Add route to get a user's permitted draftset options.
Add a route to the draftsets API which allows a user to retreive the
operations they are permitted to perform on a draftset. At present,
this simply returns the collection returned by
draftset-management/find-permitted-draftset-operations. In future, it
may be useful to return a richer structure which contains the URI
endpoints and request methods associated with the operations.