Load the user repository based on configuration.
Load the user repository from a namespace configured as an (optional)
environment variable. The DRAFTER_USER_REPO_NS environment variable
can be used to specify the namespace to locate a factory function for
creating an implementation of drafter.user.repository/UserRepository.

If no namespace is configured an in-memory repository will be created.