Add handler for restricting routes to the current draftset owner.
Many of the draftset API routes are restricted so only the current
draftset owner is authorised to call them. Add a compound
as-draftset-handler middleware which checks the request is authorised
for a known user, that the target draftset exists and that the user
is the owner of the draftset. Update the relevant route handlers to
use this middleware.