Return 422 Unprocessable Entity if live graphs do not exist.
Some of the draftsets API routes have a required graph URI parameter
which should identify a live or managed graph in the database. Add
middlewares which check the existence of incoming graphs and return
a 422 Unprocessable Entity response if the identified graph does
not exist.

Add an is-graph-live? function to draft-management which returns
whether the given graph is a public managed  graph.