Add middleware for checking the existence of required query parameters.
Add a middleware which checks the existence of require query parameters
on an incoming request. If any of the required parameters are missing,
a 422 Unprocessable Entity response is returned.

Use this middleware in the draftset SPARQL query endpoint.

Change other draftsets API routes which have optional query parameters
to return a 422 response rather than a 406 Not Acceptable response if
parameters are missing.