Add middleware to require a graph parameter for RDF triple formats.
Some routes in the draftsets API allow data to be posted in either a
quads or triples RDF format. If a triples format is provided, the
request must contain a graph parameter identifying the destination
graph. Add a middleware which checks the RDF format and checks that
a graph parameter is provided if if the format is a triples format.
If a graph is not provided, return a 422 Unprocessable Entity response.