Fix RDF output content type negotiation.
Add a drafter.requests module containing an accept function which
fetches the Accept header on a Ring request. Update the functions
in the draftsets API to use this method rather than looking in the
request map directly. Ring lower-cases HTTP header names so fix the
accept function to search for 'accept' rather than 'Accept' in the
headers.