Check user has required role before trying to claim a draftset.
Check the claiming user is in the required role before attempting to
claim a draftset. Return a forbidden response if the user is not
authorised. The claim-draftset! function still attempts to claim the
draftest atomically in the database and returns the result of the
operation. If the claim fails, this is assumed to be because of a
conflicting claim being processed between the authorisation check
and the claim being completed. In this case a 409 Conflict Detected
response is returned.