Copy statements to add or delete to a temporary file.
The add/delete statments to/from draftset jobs require the stream
containing the statements to add or delete lives longer than the
submitting request. Create a middleware to copy the incoming request
body to a tempfile and overwrite the body with the file in
subsequent handlers. This prevents the body input stream being closed
while the add or delete job is in progress.