Update 'claimable' route to include unclaimed submitted draftsets.
Update the query used by the 'claimable draftsets' route to include
draftsets submitted by the current user which have not yet been claimed.

Change the criteria for which users are able to retrieve a single
draftset. Users can now retrieve draftsets for which they are the
owner or for which they can claim in its current state.