Update swagger documentation for submit-to.
The /submit route has been renamed to /submit-to and an extra user
parameter has been added to allow draftsets to be transferred directly
to another user rather than going via the claim pool. As there are
now two methods of submitting a draftset (to a role or to a user)
the role parameter is no longer required.