Store usernames as mailto URIs.
Store the user identifiers for draftset creator/owner/submitter as a
mailto URI rather than a string containing the username. This restricts
usernames to be valid email addresses, so add a function to validate
strings contain a valid email address and check the incoming username
when creating a User record. User URIs must also be decoded when fetching
draftset information, so update get-draftset-info to extract the
usernames from the mailto URIs.