Add content-negotiation module.
Content negotiation is currently split between sparql-protocol which
defines which mime types are supported and their preference levels,
and the backends which choose how to write query results in the
negotiated format. Backends can also fail the content negotiation
if they do not support the mime type negotiated in sparql-protocol.

This approach is error-prone and requires backends to keep their
mime type mappings in sync with those supported in sparql-protocol.
The mapping defined in sparql-protocol also duplicates the information
already defined within sesame, which contains query result format
types which specify a list of supported mime types. The mapping is
incomplete which means drafter does not support all mappings supported
by sesame writers.

Add an rdf.content-negotiation module based around the sesame result
format types. It exports a negotiate function which returns a result
format object given the query type (ASK, CONSTRUCT or SELECT) and
the client's Accept request header. If negotiation fails, nil is
returned. If the negotation succeeds, backends will be responsible
for writing the query results in the given format.