Remove rdf.rewriting/result-handler-wrapper.
Drafter supports returning CONSTRUCT query results in CSV and TSV
formats, which are not supported by sesame. Sesame does support
returning SELECT query results in those formats however, so the
result-handler-wrapper was used to construct an RDFHandler wrapping
a QueryResultHandler. The wrapper converts incoming Statements
into a BindingSet containing s, p and o bindings which is then
written to the response using the inner handler.

The backend query handler implementation could be given either an
RDFHandler or QueryResultWriter by the content negotiation process
for CONSTRUCT queruies and would need to wrap any QueryResultWriter
instances in an RDFHandler before executing the query.

The content-negotiation module now registers an RDFWriter for the
CSV and TSV content types for CONSTRUCT queries. This means that
content negotiation should always return an RDFFormat for which
an RDFWriter exists if it is successful. Backends can now use the
built-in sesame functionality for constructing a result writer
from an output stream and result format without needing to do
any wrapping themselves.