Remove graph-restrictions parameter from process-sparql-query.
Graph restrictions are associated with individual SPARQL endpoints
(e.g. draftset, live, state) and are only required when preparing
queries for execution. The SparqlExecutor protocol defines the
interface execution of SPARQL queries so any graph restriction can
be encapsulated within implementations of this protocol rather than
being a parameter to process-sparql-query.

Update RewritingSesameSparqlExecutor to set the graph restriction
when preparing queries.

Add a new RestrictedExecutor record which represents an enpoint with
a graph restriction but which does not do query rewriting (e.g.
the live endpoint which is restricted to all live graphs). Update
the live and state endpoints to execute their queries through an
instance of this executor with the appropriate restriction.