Validate the union-with-live parameter on the draftset query routes.
Validate the incoming query-with-live parameter value is either
'true' or 'false' (case-insensitively) if present on the incoming
request for draftset query routes. If it is present but not one of
these strings, return a 422 Unprocessable Entity response. If it
is one of these values, parse it as a boolean and update the
union-with-live value to a boolean in the request map. If it is
not present, set the boolean value to false in the request map.