Model draftset submissions as a resource in the state graph.
At present, draftsets can be submitted directly to a user, or via
the claim mechanism through a role permitted to claim available
draftsets. Items in the unclaimed pool are identified by the
drafter:claimableBy predicate which is mapped to the string representation
of the role required to claim the draftset.

Draftset submissions to a user should also go through this claiming
process so there will be two types of submission - one for target
users and one for roles. The existing claimableBy predicate approach
is too inflexible to model two kinds of submission (short of overloading
on the value type) so add a drafter:Submission resource type to represent
a draftset submitted to a role or individual user. Draftsets can have
at most one submission in progress so the drafter:hasSubmission predicate
points from the draftset to the current submission if one exists.

Update all draftset-management functions which update or query the
drafter:claimableBy property to operate on a Submission resource
and its contained claimRole value.