Submit draftsets to users through the submission process.
Create an intermediate submission for draftsets submitted to users
rather than assigning ownership directly to the target. This means the
assigned user must explicitly claim the draftset to take ownership, and
the submitting user can re-claim the draftset while it remains unclaimed
by the target.

Update the queries for finding all draftsets claimable by a user to
include those for which there is a submission with the user as the
target.