Add silent parameter to 'delete graph' operation.
Add a silent parameter to the 'delete graph' operation within a draftset.
If this flag is present a client error will not be returned if the
deleting graph is not managed. Instead, no action should be taken and
a 200 response returned.

Add a return type of the new draftset information to the spec.