Remove /draftsets/claimable route and add filter to /draftsets route.
The /draftsets route currently returns all draftsets visible to the
requesting user - these are the draftsets owned by the user and those
which they can claim. The /claimable route returns just the draftsets
the requesting user does not own but can claim. There is no route which
just returns the draftsets owned by the current user.

Add an 'include' enum parameter with values 'owned', 'claimable' and
'all' to the /draftsets route which allows the reqesting user to
specify which subset of visible draftests they want to be returned.

Remove the /draftsets/claimable route since it is now equivalent to
/draftsets?include=claimable.