Refactor 'get draftsets' queries.
The 'get draftsets for user' queries can return a draftset multiple
times if the user is eligible to claim the draftset in multiple ways.
For example, if a user submits to a role they are in they will be
eligible to claim the draftset both as the submitter and as a member of
the claim role.

Refactor the 'get draftsets' queries to use a sub-select which returns
the unique collection of draftset URIs the current user is eligible
to claim. Embed this sub-query into one which returns either the
properties of the draftset or the draftset graph mappings.

Extract each of the draftset matching clauses (e.g. user in claim
role, user is owner etc.) into their own functions and pass a
list of the clauses to build the resulting query to find matching
draftsets.

Remove duplicates from 'all draftsets' query.

Refactor draftsets queries and remove unused function.

Refactor graph mappings queries.

Use properties and graph mappings query for fetching single draftset.

Move draftset type constraint into properties and graphs queries.

Extract draftset visibility clauses.