Require mongodb to be explicitly configured.
Remove the hard-coded connection and database settings from the
Mongodb user repository and fetch them from the environment. The
user database name must always be configured, while the server location
can optionally be speicified with host and port settings. If one of the
host and port is provided the other is required.

Fix the mongodb document to drafter user mapping so the
encrypted_password field is used for the password hash instead of
the api_key_digest field.